FROMLIST: media: iris: Allow stop on firmware only if start was issued.
For HFI Gen1, the instances substate is changed to LOAD_RESOURCES only
when a START command is issues to the firmware. If STOP is called
without a prior START, the firmware may reject the command and throw
some erros.
Handle this by adding a substate check before issuing STOP command to
the firmware.

Link: https://lore.kernel.org/linux-media/20250704-iris-video-encoder-v1-9-b6ce24e273cf@quicinc.com/
Fixes: 11712ce70f8e ("media: iris: implement vb2 streaming ops")
Signed-off-by: Dikshita Agarwal <quic_dikshita@quicinc.com>